datatype consistency fix

+++ algiers-lib/models/landmark.py
@@ -53,12 +53,12 @@ class TimeSlot:
                 f"open_time ({self.open_time}) must be strictly less than "
                 f"close_time ({self.close_time}).")
 
-    def contains(self, arrival: int, duration: int) -> bool:
+    def contains(self, arrival: float, duration: float) -> bool:
         """Check if a visit starting at arrival time with given duration fits within the slot.
 
         Args:
-            arrival (int): Arrival time in minutes.
-            duration (int): Visit duration in minutes.
+            arrival (float): Arrival time in minutes.
+            duration (float): Visit duration in minutes.
 
         Returns:
             bool: True if the visit fits, False otherwise.
